chore: extract l1 logic
